requirements.txt
# Core Flask dependencies
Flask==3.1.1
Flask-Session==0.8.0

# Production server
gunicorn==23.0.0

# Environment and configuration
python-dotenv==1.1.1

# AI/ML dependencies
openai==0.28.0

# HTTP and networking
requests==2.32.3

# Data processing
numpy==2.2.5

# Development and testing (optional in production)
pytest==8.3.3
pytest-flask==1.3.0

# Security
Werkzeug==3.1.3
